Move mat-custom-icon service to seperate folder
