+++ src/app/app.component.spec.ts
@@ -7,7 +7,7 @@ import { RouterStateService } from './state/router-state/router-state.service';
 import { StoreModule } from '@ngrx/store';
 import { reducers } from './state';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { MatCustomIconService } from './shared/services/mat-custom-icon.service';
+import { MatCustomIconService } from './shared/services/mat-custom-icon/mat-custom-icon.service';
 import { MatIconModule } from '@angular/material';
 import { HttpClientModule } from '@angular/common/http';
