+++ src/app/shared/services/mat-custom-icon/mat-custom-icon.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { MatIconRegistry } from '@angular/material';
 import { DomSanitizer } from '@angular/platform-browser';
-import { customIcons } from '../data/custom-icons';
+import { customIcons } from '../../data/custom-icons';
 
 @Injectable({
   providedIn: 'root'
